          \name{CoxRidge internal}
          \alias{sumevents}
          \alias{sumeventsf}
          \title{Internal \code{CoxRidge} functions.}
          \description{
            Internal \code{CoxRidge} functions.}
  
          \details{
           These are not to be called by the user}


          \keyword{sumevents,sumevents}
