\name{CoxRidge-package}
\alias{CoxRidge-package}
\alias{CoxRidge}
\docType{package}
\title{Fit penalized Cox models with fixed and time varying effects of the covariates.}
\description{
Fits penalized Cox models using ridge penalties and a REML-type algorithm for optimization. The methods can be applied also to non-proportional hazards models where some or all of the covariates can be modelled with time varying effects. 
}
\details{
\tabular{ll}{
Package: \tab CoxRidge\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2013-06-20\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Aris Perperoglou <aperpe@essex.ac.uk>
Maintainer: Aris Perperoglou <aperpe@essex.ac.uk>
}
\references{
Perperoglou A.(2013)\emph{Cox models with dynamic ridge penalties on
time varying effects of the covariates}. Statistics in Medicine, to appear}

\keyword{ package }
